Add new func shape_create_polygon().
